Hostile Lockdown Optimizations + Bolt lights
- Shock-bolting all doors has been made more efficient on the server.
This greatly reduces lag potential for AIs using it as a door crush.

- Introduced a small cooldown (three BYOND seconds) in a further
optimization to remove a critical exploit.

- Due to the way the new code works, the module no longer triggers
firealarms but closes the locks directly. <b>To compensate for this,
bolt lights on airlocks will remain on.</b> So, instead of red fire
alarm lights, all the airlocks will have bolt lights.