Blobs can now occur with 17 players or more.
Blobs with 17-21 players ready get a 1.75 resource rate instead of the usual 3.
Blobs with 22-29 players ready get a 2.3 resource rate instead of the usual 3.
30+ player blobs keep their resource rate of 3.

Blobs also have to expand a bit more to win now. This should discourage hard-to-reach-without-AI locations such as xenobio or the AI sat which are a problem on lower pop rounds.